+++ bika/lims/content/analysisservice.py
@@ -32,7 +32,6 @@ from bika.lims.content.abstractbaseanalysis import AbstractBaseAnalysis
 from bika.lims.content.abstractbaseanalysis import schema
 from bika.lims.interfaces import IAnalysisService
 from bika.lims.interfaces import IDeactivable
-from bika.lims.interfaces import IHaveIdentifiers
 from bika.lims.utils import to_utf8 as _c
 from magnitude import mg
 from Products.Archetypes.public import BooleanField
@@ -419,7 +418,7 @@ schema.moveField('InterimFields', after='Calculation')
 
 
 class AnalysisService(AbstractBaseAnalysis):
-    implements(IAnalysisService, IHaveIdentifiers, IDeactivable)
+    implements(IAnalysisService, IDeactivable)
     security = ClassSecurityInfo()
     schema = schema
     displayContentsTab = False