+++ bika/lims/upgrade/v01_03_002.py
@@ -57,6 +57,9 @@ def upgrade(tool):
     # Allow to detach a partition from its primary sample (#1420)
     update_partitions_role_mappings(portal)
 
+    # Remove Identifiers
+    remove_identifiers(portal)
+
     # Unindex stale catalog brains from the auditlog_catalog
     # https://github.com/senaite/senaite.core/issues/1438
     unindex_orphaned_brains_in_auditlog_catalog(portal)
@@ -129,3 +132,39 @@ def update_partitions_role_mappings(portal):
         partition.reindexObjectSecurity()
 
     logger.info("Updating role mappings of partitions [DONE]")
+
+
+def remove_identifiers(portal):
+    """Remove Identifiers from the portal
+    """
+    # 1. Remove the identifiers and identifier types
+    logger.info("Removing identifiers ...")
+
+    setup = portal.bika_setup
+    try:
+        # we use _delOb because manage_delObjects raises an unauthorized here
+        it = setup["bika_identifiertypes"]
+        for i in it.objectValues():
+            i.unindexObject()
+        it.unindexObject()
+        setup._delOb("bika_identifiertypes")
+    except KeyError:
+        pass
+
+    # 2. Remove controlpanel configlet
+    cp = portal.portal_controlpanel
+    cp.unregisterConfiglet("bika_identifiertypes")
+
+    # 3. Remove catalog indexes
+    for cat in ["portal_catalog", "bika_catalog", "bika_setup_catalog"]:
+        tool = portal[cat]
+        if "Identifiers" in tool.indexes():
+            tool.manage_delIndex("Identifiers")
+
+    # 4. Remove type registration
+    pt = portal.portal_types
+    for t in ["IdentifierType", "IdentifierTypes"]:
+        if t in pt.objectIds():
+            pt.manage_delObjects(t)
+
+    logger.info("Removing identifiers [DONE]")